Clean up LibraryItem (#9072)
* Move LibraryItem vars to constructor vals

* Convert LibraryItem to data class

Remove redundant equals and hashCode

* Remove unused LibraryItem.displayMode

* Simplify LibraryItem.matches()

* Align types in LibraryItem and LibraryBadges

* fixup! Simplify LibraryItem.matches()

+++ app/src/main/java/eu/kanade/presentation/library/components/LibraryBadges.kt
@@ -7,7 +7,7 @@ import eu.kanade.presentation.components.Badge
 import eu.kanade.tachiyomi.R
 
 @Composable
-fun DownloadsBadge(count: Int) {
+fun DownloadsBadge(count: Long) {
     if (count > 0) {
         Badge(
             text = "$count",
@@ -18,7 +18,7 @@ fun DownloadsBadge(count: Int) {
 }
 
 @Composable
-fun UnreadBadge(count: Int) {
+fun UnreadBadge(count: Long) {
     if (count > 0) {
         Badge(text = "$count")
     }
